refactor: AuthenticationConverter 상속 -> 합성
