+++ src/main/java/nextstep/auth/authentication/AuthenticationConverter.java
@@ -10,9 +10,9 @@ import java.io.IOException;
 public interface AuthenticationConverter {
     AuthenticationToken convert(HttpServletRequest request) throws IOException;
 
-    default Authentication authenticate(AuthenticationToken token, UserDetailsServiceStrategy userDetailsService) {
-        String principal = token.getPrincipal();
-        final LoginMember userDetails = userDetailsService.authenticate(principal, token);
-        return new Authentication(userDetails);
-    }
+//    default Authentication authenticate(AuthenticationToken token, UserDetailsServiceStrategy userDetailsService) {
+//        String principal = token.getPrincipal();
+//        final LoginMember userDetails = userDetailsService.authenticate(principal, token);
+//        return new Authentication(userDetails);
+//    }
 }
