+++ src/main/java/nextstep/auth/authentication/TokenAuthenticationConverter.java
@@ -1,0 +1,15 @@
+package nextstep.auth.authentication;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import nextstep.auth.token.TokenRequest;
+
+import javax.servlet.http.HttpServletRequest;
+import java.io.IOException;
+
+public class TokenAuthenticationConverter implements AuthenticationConverter{
+    @Override
+    public AuthenticationToken convert(HttpServletRequest request) throws IOException {
+        final TokenRequest tokenRequest = new ObjectMapper().readValue(request.getInputStream(), TokenRequest.class);
+        return new AuthenticationToken(tokenRequest.getEmail(), tokenRequest.getPassword());
+    }
+}
